refactor(lotto): 마무리하지 못한 리팩토링
