+++ src/main/java/lotto/service/Calculated.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.service;
 
 import java.util.Map;
 import lotto.rank.Ranking;
